finish hw5

+++ views/posts.py
@@ -1,10 +1,12 @@
 from flask import Response, request
 from flask_restful import Resource
-from models import Post, db
-
-from views import security, get_authorized_user_ids
-
+from models import Post, User, db
+from . import can_view_post, get_authorized_user_ids
 import json
+from sqlalchemy import and_
+from my_decorators import secure_bookmark, \
+    handle_db_insert_error, check_ownership_of_post, is_valid_int, id_is_integer_or_400_error, secure_post
+import flask_jwt_extended
 
 def get_path():
     return request.host_url + 'api/posts/'
@@ -14,9 +16,10 @@ class PostListEndpoint(Resource):
     def __init__(self, current_user):
         self.current_user = current_user
 
+    @flask_jwt_extended.jwt_required()
     def get(self):
-        ids = get_authorized_user_ids(self.current_user)
-        posts = Post.query.filter(Post.user_id.in_(ids))
+        ids_for_me_and_my_friends = get_authorized_user_ids(self.current_user)
+        posts = Post.query.filter(Post.user_id.in_(ids_for_me_and_my_friends))
         limit = request.args.get('limit')
         if limit:
             try:
@@ -28,19 +31,18 @@ class PostListEndpoint(Resource):
         else:
             limit = 10
         posts = posts.order_by(Post.pub_date.desc()).limit(limit)
-
-        # note: if you pass the current user into to_dict(), it will tell you
-        # whether or not the current user liked and/or bookmarked any of the posts
         data = [
-            item.to_dict(user=self.current_user) for item in posts.all()
+            item.to_dict(user=self.current_user) for item in posts
         ]
         return Response(json.dumps(data), mimetype="application/json", status=200)
 
+    @flask_jwt_extended.jwt_required()
+    @handle_db_insert_error
     def post(self):
         body = request.get_json()
+        if not body:
+            return Response(json.dumps({'message': 'Bad data'}), mimetype="application/json", status=400)
         image_url = body.get('image_url')
-        if image_url is None:
-            return Response(json.dumps({'message': '"image_url" is required.'}), mimetype="application/json", status=400)
         caption = body.get('caption')
         alt_text = body.get('alt_text')
         user_id = self.current_user.id # id of the user who is logged in
@@ -55,49 +57,66 @@ class PostDetailEndpoint(Resource):
 
     def __init__(self, current_user):
         self.current_user = current_user
-        
-    
-    @security.id_is_valid
-    @security.user_can_edit_post
+
+    @flask_jwt_extended.jwt_required()  
+    @id_is_integer_or_400_error
+    @handle_db_insert_error
+    @check_ownership_of_post
     def patch(self, id):
         post = Post.query.get(id)
-        body = request.get_json()
 
-        post.image_url = body.get('image_url') or body.get('image_url')
-        if body.get('caption') is not None:
-            post.caption = body.get('caption')
-        if body.get('alt_text') is not None:
-            post.alt_text = body.get('alt_text')
+        # a user can only edit their own post:
+        if not post or post.user_id != self.current_user.id:
+            return Response(json.dumps({'message': 'Post does not exist'}), mimetype="application/json", status=404)
+       
+
+        body = request.get_json()
+        post.image_url = body.get('image_url') or post.image_url
+        post.caption = body.get('caption') or post.caption
+        post.alt_text = body.get('alt_text') or post.alt_text
         
         # commit changes:
         db.session.commit()        
         return Response(json.dumps(post.to_dict(user=self.current_user)), mimetype="application/json", status=200)
     
-    @security.id_is_valid
-    @security.user_can_edit_post
+    @flask_jwt_extended.jwt_required()
+    @id_is_integer_or_400_error
+    @check_ownership_of_post
     def delete(self, id):
+
+        # a user can only delete their own post:
+        post = Post.query.get(id)
+        if not post or post.user_id != self.current_user.id:
+            return Response(json.dumps({'message': 'Post does not exist'}), mimetype="application/json", status=404)
+       
+
         Post.query.filter_by(id=id).delete()
         db.session.commit()
-
         serialized_data = {
             'message': 'Post {0} successfully deleted.'.format(id)
         }
         return Response(json.dumps(serialized_data), mimetype="application/json", status=200)
-
-    @security.id_is_valid
-    @security.user_can_view_post
+    
+    @flask_jwt_extended.jwt_required()
+    @id_is_integer_or_400_error
+    @secure_post
     def get(self, id):
         post = Post.query.get(id)
+
+        # if the user is not allowed to see the post or if the post does not exist, return 404:
+        if not post or not can_view_post(post.id, self.current_user):
+            return Response(json.dumps({'message': 'Post does not exist'}), mimetype="application/json", status=404)
+        
         return Response(json.dumps(post.to_dict(user=self.current_user)), mimetype="application/json", status=200)
 
 def initialize_routes(api):
     api.add_resource(
         PostListEndpoint, 
         '/api/posts', '/api/posts/', 
-        resource_class_kwargs={'current_user': api.app.current_user}
+        resource_class_kwargs={'current_user': flask_jwt_extended.current_user}
     )
     api.add_resource(
         PostDetailEndpoint, 
         '/api/posts/<id>', '/api/posts/<id>/',
-        resource_class_kwargs={'current_user': api.app.current_user}
+        resource_class_kwargs={'current_user': flask_jwt_extended.current_user}
     )
